+++ src/content/scripting/script_names.h
@@ -51,6 +51,7 @@ const static auto res_names = std::map<ResourceAttribute, std::string> {
     std::pair(ResourceAttribute::VIDEOCODEC, "R_VIDEOCODEC"),
     std::pair(ResourceAttribute::ORIENTATION, "R_ORIENTATION"),
     std::pair(ResourceAttribute::PIXELFORMAT, "R_PIXELFORMAT"),
+    std::pair(ResourceAttribute::LYRICS, "R_LYRICS"),
 };
 
 const static auto mt_names = std::map<MetadataFields, std::string_view> {